+++ src/steamship/agents/react/react.py
@@ -77,7 +77,7 @@ New input: {input}
             tool_names=tool_names,
             scratchpad=scratchpad,
             chat_history=context.chat_history.messages_as_string(
-                memory_window_strategy=self.memory_window_strategy
+                memory_window_strategy=self.memory_strategy
             ),
         )
