+++ src/steamship/agents/schema/agent.py
@@ -3,10 +3,11 @@ from typing import List
 
 from pydantic import BaseModel
 
+from steamship.agents.memory.chathistory import MemoryStrategy
 from steamship.agents.schema.action import Action
 from steamship.agents.schema.context import AgentContext
-from steamship.agents.schema.conversation_memory import MemoryWindowStrategy, NoMemory
 from steamship.agents.schema.llm import LLM
+from steamship.agents.schema.memory_strategies import NoMemory
 from steamship.agents.schema.output_parser import OutputParser
 from steamship.agents.schema.tool import Tool
 
@@ -21,7 +22,7 @@ class Agent(BaseModel, ABC):
     tools: List[Tool]
     """Tools that can be used by the Agent in selecting the next Action."""
 
-    memory_window_strategy: MemoryWindowStrategy = NoMemory()
+    memory_strategy: MemoryStrategy = NoMemory()
     """Conversation memory to use when running agent."""
 
     @abstractmethod
